refactor: combine property service and propertycontrollers to one class
